Refactor wizard state into store and add preview panel
