+++ pages/creation.vue
@@ -1,12 +1,15 @@
-## pages/creation.vue (Module_Bonome — UI principale)
-```vue
 <template>
-  <section class="carte">
-    <h1 class="h1">Création de personnage (Module_Bonome)</h1>
-    <BonomeWizard />
-  </section>
+  <div class="space-y-6">
+    <section class="carte">
+      <BonomeWizard />
+    </section>
+    <section class="carte">
+      <BonomePreviewPanel />
+    </section>
+  </div>
 </template>
 
 <script setup lang="ts">
-import BonomeWizard from '@/components/BonomeWizard.vue'
+import BonomeWizard from '@/components/BonomeWizard.vue';
+import BonomePreviewPanel from '@/components/BonomePreviewPanel.vue';
 </script>
